checking if forces.csv is existing

+++ start_plotting.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/bash
 
-rm forces.csv
+FILE=forces.csv
+if [ -f "$FILE" ]; then
+	rm forces.csv
+fi
 python3 Service_Force_Messurement.py &
 python3 plotForces.py
